CAPPI can now draw in either ray mode(beams are wedge shaped) or strip mode (beams are strips). Fixed timetag calculation in CappiReader which was fouling up delta time calculations.
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@2655 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ src/eldoracappi/CappiReader.h
@@ -16,7 +16,7 @@ class CappiReader {
     bool openFile(std::string inputFile);
     void closeFile();
     
-    bool read(unsigned long index, std::vector<double> &p, int &prodType, long long &timeTag, StrMapDouble &hskpMap);
+    bool read(unsigned long index, std::vector<double> &p, int &prodType, double &timeTag, StrMapDouble &hskpMap);
 
     bool findLastRecord(unsigned long &index);
