+++ src/eldoracappi/EldoraCappiMain.cpp
@@ -103,9 +103,9 @@ int main(
     EldoraCappi cappi(inputFile, title, dialog);
 
     QObject::connect(&cappi,
-         SIGNAL(newProductSignal(DoubleVec,int, qlonglong, StrMapDouble)),
+         SIGNAL(newProductSignal(DoubleVec,int, double, StrMapDouble)),
                             &cappi,
-           SLOT(productSlot(DoubleVec,int, qlonglong, StrMapDouble)) );
+           SLOT(productSlot(DoubleVec,int, double, StrMapDouble)) );
     
 
     // if we don't show() the dialog, nothing appears!
